requirements.txt
Telethon~=1.24.0
PyYAML~=6.0